removing all the unecessary `:page/:foo` legacy syntax tests
