+++ test/test.js
@@ -1,3 +1,2 @@
 require("./route-test");
-require("./route-curlies-syntax-test");
 require("./route-define-test");
